Change codes

+++ rust_by_example/15/rx_15_4/src/main.rs
@@ -1,18 +1,18 @@
 struct Fibonacci {
-    curr: u32,
-    next: u32,
+    curr: u64,
+    next: u64,
 }
 
 // `Iterator`を`Fibonacci`に対して実装する。
 // `Iterator`トレイトは次(`next`)の要素を取得するメソッドの定義だけを要求する。
 impl Iterator for Fibonacci {
-    type Item = u32;
+    type Item = u64;
 
     // ここではイテレーションの流れを`.curr`と`.next`を使用して定義している。
     // 返り値の型は`Option<T>`で、これは:
     //     * `Iterator`が終了した時は`None`を返し、
     //     * そうでなければ`Some`でラップされた値を返す。
-    fn next(&mut self) -> Option<u32> {
+    fn next(&mut self) -> Option<u64> {
         let new_next = self.curr + self.next;
 
         self.curr = self.next;
